Extract app reducer intents function
